Move Build tree
